push

+++ WatchinWeebsBot/Bot.cs
@@ -76,13 +76,20 @@ namespace WatchinWeebsBot
 
         async Task RandomlyPostTheFunny()
         {
-            Console.WriteLine("posting the funny");
-            Random rnd = new Random();
-            string rndpath = Directory.GetFiles("memes").ElementAt(rnd.Next(1, Directory.GetFiles("memes").Length));
-            await Client.GetGuildAsync(691036170238427186).Result.GetChannel(729099294669275228).SendFileAsync(rndpath, "Here is your meme!");
-            await Task.Delay(3600000);
-            await RandomlyPostTheFunny();
-            //await Client.GetGuildAsync(691036170238427186).Result.GetChannel(729099294669275228).SendFileAsync(rndpath);
+            try
+            {
+                Console.WriteLine("posting the funny");
+                Random rnd = new Random();
+                string rndpath = Directory.GetFiles("memes").ElementAt(rnd.Next(1, Directory.GetFiles("memes").Length));
+                await Client.GetGuildAsync(691036170238427186).Result.GetChannel(729099294669275228).SendFileAsync(rndpath, "Here is your meme!");
+                await Task.Delay(3600000);
+                await RandomlyPostTheFunny();
+                //await Client.GetGuildAsync(691036170238427186).Result.GetChannel(729099294669275228).SendFileAsync(rndpath);
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+            }
         }
 
         async Task PostTheFunny(MessageCreateEventArgs a)
@@ -167,14 +174,37 @@ namespace WatchinWeebsBot
                 }
             }
             */
-                if (e.Message.Content.ToLower().Contains("hello nezuko"))
+                if (e.Message.Content.ToLower().Contains("hello nezuko") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     await e.Channel.SendMessageAsync("Hello!");
                 }
-                if (e.Message.Content.ToLower().Contains("!nezukoupdatememe") && CheckIfCoolPerson(e))
+                if (e.Message.Content.ToLower().Contains("!nezukoupdatememe") && CheckIfCoolPerson(e) && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     await e.Channel.SendMessageAsync("Updating my memes!");
                     await UpdateTheFunny(e);
+                }
+                if (e.Message.Content.ToLower().Contains("!commands"))
+                {
+                    await e.Channel.SendMessageAsync("Here are the commands: \n" +
+                        "hello nezuko \n" +
+                        "!nezukoupdateme \n" +
+                        "!postthefunny \n" +
+                        "!restart \n" +
+                        "!quit \n" +
+                        "!nezban \n" +
+                        "!showmeallcool \n" +
+                        "!smsg \n" +
+                        "nezuko am i cool \n" +
+                        "nezuko is he cool \n" +
+                        "!delmsg \n" +
+                        "make channel nezuko \n" +
+                        "delete channel nezuko \n" +
+                        "!addcool \n" +
+                        "!addcool \n" +
+                        "!removecool \n" +
+                        "!trace \n" +
+                        "!untrace \n");
+
                 }
                 /*
                 if(e.Message.Content.ToLower().Contains("nezuko make role"))
@@ -184,12 +214,12 @@ namespace WatchinWeebsBot
                     await e.Channel.SendMessageAsync("Role has been made!");
                 }
                 */
-                if (e.Message.Author.Id.ToString() == "453826077442179072" || e.Message.Author.Id.ToString() == "238327938859270145")
+                if (e.Message.Author.Id.ToString() == "453826077442179072" || e.Message.Author.Id.ToString() == "238327938859270145" && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
-                    await e.Message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":sweat_drops:"));
                     await e.Message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":confounded:"));
+                    await e.Message.CreateReactionAsync(DiscordEmoji.FromName(Client, ":sweat_drops:"));
                 }
-                if (e.Message.Content.ToLower().Contains("klives"))
+                if (e.Message.Content.ToLower().Contains("klives") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     try
                     {
@@ -200,11 +230,11 @@ namespace WatchinWeebsBot
                         Console.WriteLine(ex.Message);
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!postthefunny"))
+                if (e.Message.Content.ToLower().Contains("!postthefunny") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     await PostTheFunny(e);
                 }
-                if (e.Message.Content.Contains("!restart"))
+                if (e.Message.Content.Contains("!restart") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -217,7 +247,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!nezban"))
+                if (e.Message.Content.ToLower().Contains("!nezban") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -225,7 +255,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync(e.Message.MentionedUsers.ElementAtOrDefault(0).Username);
                     }
                 }
-                if (e.Message.Content.Contains("!quit"))
+                if (e.Message.Content.Contains("!quit") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (e.Message.Author.Id.ToString() == "227462990293762049")
                     {
@@ -237,14 +267,14 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only Klives can execute this!.");
                     }
                 }
-                if (e.Message.Content.Contains("!showmeallcool"))
+                if (e.Message.Content.Contains("!showmeallcool") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     foreach(string item in important)
                     {
-                        e.Channel.SendMessageAsync(item);
+                        await e.Channel.SendMessageAsync(item);
                     }
                 }
-                if (e.Message.Content.Contains("!smsg"))
+                if (e.Message.Content.Contains("!smsg") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -260,7 +290,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("nezuko am i cool"))
+                if (e.Message.Content.ToLower().Contains("nezuko am i cool") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -272,7 +302,7 @@ namespace WatchinWeebsBot
 
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("nezuko is he cool"))
+                if (e.Message.Content.ToLower().Contains("nezuko is he cool") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (important.Contains(e.Message.MentionedUsers.ElementAtOrDefault(0).Id.ToString()))
                     {
@@ -283,7 +313,7 @@ namespace WatchinWeebsBot
                         await e.Channel.SendMessageAsync("No!");
                     }
                 }
-                if (e.Message.Content.Contains("!delmsg"))
+                if (e.Message.Content.Contains("!delmsg") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     await e.Channel.GetMessageAsync(Convert.ToUInt64(e.Message.Content.Replace("!delmsg ", string.Empty))).Result.DeleteAsync();
                     await e.Message.DeleteAsync();
@@ -301,7 +331,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("delete channel nezuko"))
+                if (e.Message.Content.ToLower().Contains("delete channel nezuko") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -313,7 +343,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("lock this chat"))
+                if (e.Message.Content.ToLower().Contains("lock this chat") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (e.Message.Author.Id.ToString() == "227462990293762049")
                     {
@@ -325,12 +355,12 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!addcool"))
+                if (e.Message.Content.ToLower().Contains("!addcool") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     var guy = e.Message.MentionedUsers.ElementAtOrDefault(0).Id.ToString();
                     if (CheckIfCoolPerson(e))
                     {
-                        if (guy == "257204702528274433")
+                        if (guy == "257204702528274433" && e.Message.Author.Id.ToString() == "737326568271118457")
                         {
                             await e.Message.Channel.SendMessageAsync("Stefo isn't going to get cool person.");
                             await e.Guild.GetMemberAsync(227462990293762049).Result.SendMessageAsync(e.Message.Author.Username + " tried giving Stefo cool role. Bad boy.");
@@ -346,7 +376,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!removecool"))
+                if (e.Message.Content.ToLower().Contains("!removecool") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -367,7 +397,7 @@ namespace WatchinWeebsBot
                         await e.Message.Channel.SendMessageAsync("Only cool people can execute this!.");
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("unlock this chat"))
+                if (e.Message.Content.ToLower().Contains("unlock this chat") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (e.Message.Author.Id.ToString() == "227462990293762049")
                     {
@@ -380,7 +410,7 @@ namespace WatchinWeebsBot
                         }
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!trace"))
+                if (e.Message.Content.ToLower().Contains("!trace") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -388,7 +418,7 @@ namespace WatchinWeebsBot
                         await e.Message.DeleteAsync();
                     }
                 }
-                if (e.Message.Content.ToLower().Contains("!untrace"))
+                if (e.Message.Content.ToLower().Contains("!untrace") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     if (CheckIfCoolPerson(e))
                     {
@@ -396,7 +426,7 @@ namespace WatchinWeebsBot
                         await e.Message.DeleteAsync();
                     }
                 }
-                if (tracing && e.Message.Author.Id.ToString() == "227462990293762049" && !e.Message.Content.Contains("!trace") && !e.Message.Content.Contains("!untrace"))
+                if (tracing && e.Message.Author.Id.ToString() == "227462990293762049" && !e.Message.Content.Contains("!trace") && !e.Message.Content.Contains("!untrace") && e.Message.Author.Id.ToString() == "737326568271118457")
                 {
                     await e.Message.Channel.SendMessageAsync(e.Message.Content);
                     await e.Message.DeleteAsync();
